refactor(e2e): extract common scaffold + federation + reputation modules (sub-phases 1-3/7)
Sub-phase 1 — common scaffold:
  tests/e2e/common/mod.rs: governance_fixtures, EnvVarGuard, SingleI32

Sub-phase 2 — federation modules:
  tests/e2e/federation_inbound_a.rs (2 tests)
  tests/e2e/federation_inbound_b.rs (6 tests)
  tests/e2e/federation_inbound_e.rs (1 test)
  All use #[path] + crate::common::governance_fixtures (not super::*)

Sub-phase 3 — reputation modules:
  tests/e2e/reputation_rt_r3.rs (12 tests)
  tests/e2e/reputation_rt_r4.rs (7 tests)

GOLDEN_INVARIANT: checkpoint #2 (bx6witdrp) confirmed 130 passed, 5 skipped.
Compile gate: SUB2_COMPILE_EXIT_0 (cargo-test.bat --no-run --workspace --features full).
