+++ crates/server/tests/e2e.rs
@@ -108,818 +108,27 @@ async fn template_dump_capture() -> lemmy_utils::error::LemmyResult<()> {
   Ok(())
 }
 
-struct EnvVarGuard {
-  key: &'static str,
-  prev: Option<String>,
-}
-
-impl EnvVarGuard {
-  fn set(key: &'static str, value: &str) -> Self {
-    let prev = std::env::var(key).ok();
-    // SAFETY: tests run with --test-threads=1; no concurrent env mutation.
-    unsafe {
-      std::env::set_var(key, value);
-    }
-    Self { key, prev }
-  }
-}
-
-impl Drop for EnvVarGuard {
-  fn drop(&mut self) {
-    // SAFETY: same justification — single-threaded test runner.
-    unsafe {
-      match &self.prev {
-        Some(prev) => std::env::set_var(self.key, prev),
-        None => std::env::remove_var(self.key),
-      }
-    }
-  }
-}
+// Shared cross-domain test scaffold (Phase 6 decomposition, sub-phase 1):
+// the testcontainer/Postgres harness, the process-env RAII guard, and the
+// single-column row shape now live in `tests/e2e/common/mod.rs`. The `use`
+// re-exports below keep the ~291 bare `governance_fixtures::` /
+// `EnvVarGuard` / `SingleI32` references in the test bodies unchanged.
+//
+// `#[path]` is mandatory: this file is `tests/e2e.rs`, an auto-discovered
+// integration-test crate root, so a bare `mod common;` resolves to
+// `tests/common/mod.rs` (the sibling dir), NOT `tests/e2e/common/`. The
+// explicit path keeps all split modules under `tests/e2e/` while preserving
+// the single-test-binary contract — files placed directly in `tests/` would
+// each become a SEPARATE test binary, which we must avoid.
+#[path = "e2e/common/mod.rs"]
+mod common;
+use common::governance_fixtures;
+#[allow(unused_imports)]
+use common::{EnvVarGuard, SingleI32};
 
 // ============================================================================
 // Phase 1 — governance schema + hash-chain trigger smoke tests
 // ============================================================================
-
-mod governance_fixtures {
-  //! **Process-env safety constraint:** every test in this module mutates
-  //! process-wide env vars (e.g. `LEMMY_DATABASE_URL`, `BREHON_DISABLE_*`).
-  //! Safety of those mutations is contingent on the Cargo runner flag
-  //! `--test-threads=1`. Running these tests with concurrent threads is
-  //! undefined behaviour and is forbidden — see the `EnvVarGuard` RAII guard.
-  use super::EnvVarGuard;
-  use actix_web::web::Data;
-  use diesel::{Connection as _, PgConnection, RunQueryDsl, connection::SimpleConnection};
-  use diesel_migrations::{EmbeddedMigrations, MigrationHarness, embed_migrations};
-  use lemmy_api_utils::{context::LemmyContext, request::client_builder};
-  use lemmy_db_schema::source::{
-    community::{Community, CommunityInsertForm},
-    local_user::{LocalUser, LocalUserInsertForm},
-    person::{Person, PersonInsertForm},
-    secret::Secret,
-  };
-  use lemmy_db_schema_file::{InstanceId, PersonId};
-  use lemmy_db_views_local_user::LocalUserView;
-  use lemmy_diesel_utils::{
-    connection::{ActualDbPool, build_db_pool_for_tests},
-    traits::Crud,
-  };
-  use lemmy_utils::{error::LemmyResult, rate_limit::RateLimit, settings::SETTINGS};
-  use reqwest_middleware::ClientBuilder;
-  use std::error::Error;
-
-  /// Migrations are embedded at compile time from the repo-root `migrations/`
-  /// directory. Path is relative to `CARGO_MANIFEST_DIR` (here,
-  /// `crates/server`), so `../../migrations` resolves to repo-root
-  /// `migrations/`.
-  const MIGRATIONS: EmbeddedMigrations = embed_migrations!("../../migrations");
-
-  /// Apply the full Lemmy + governance schema to a fresh container:
-  ///   1. Acquire `pg_advisory_lock(0)` to bypass the `forbid_diesel_cli`
-  ///      trigger installed in migration `2025-08-01-000017` (it rejects any
-  ///      insert into `__diesel_schema_migrations` that doesn't hold the
-  ///      lock).
-  ///   2. Run every embedded migration via diesel's standard
-  ///      `MigrationHarness::run_pending_migrations`.
-  ///   3. Rebuild the `r` schema (replaceable-schema layer) by inlining
-  ///      `utils.sql` + `triggers.sql` — the same two files that
-  ///      `lemmy_diesel_utils::schema_setup` loads via `include_str!`. This
-  ///      installs the governance hash-chain and append-only triggers on
-  ///      `public.governance_log`.
-  ///
-  /// **Tier 3 fast path:** when `start_postgres` has already restored the
-  /// template dump into this container, the schema is fully populated and
-  /// steps 1-3 would re-do work the dump already captured. The sentinel
-  /// short-circuits to `Ok(())` so existing call sites that pair
-  /// `start_postgres()` + `apply_all_schema(&mut conn)` get the speedup
-  /// transparently.
-  ///
-  /// **Sentinel correctness (CR finding #6):** `governance_log` (in `public`)
-  /// is introduced by a mid-set migration, so a partial-init DB could have
-  /// `governance_log` but be missing later migrations or the `r` schema —
-  /// short-circuiting on `governance_log` alone would mask that broken
-  /// state. The fix is to require BOTH `public.governance_log` (proves
-  /// all migrations ran through that point) AND `r.parent_comment_ids`
-  /// (proves the last replaceable-schema rebuild completed — `triggers.sql`
-  /// is the last include in `apply_all_schema_legacy`). If either is
-  /// missing, fall through to the legacy bootstrap rather than risking
-  /// silent partial-init.
-  ///
-  /// To force the legacy path (cold migrations) for debugging, set
-  /// `BREHON_E2E_NO_TEMPLATE=1` before the test run — `start_postgres`
-  /// then skips the restore and the sentinel here misses.
-  /// Two-signal schema sentinel (CR finding #6 + #16). Returns true
-  /// only when BOTH `public.governance_log` (proves all migrations
-  /// ran through that point) AND `r.parent_comment_ids` (proves the
-  /// LAST step of `apply_all_schema_legacy` — `triggers.sql` —
-  /// completed) are present. Either alone indicates a partial-init
-  /// DB that must be re-bootstrapped, not skipped.
-  ///
-  /// Single source of truth — used by `apply_all_schema` for the
-  /// fast-path short-circuit AND by `postgres_container_boots` for
-  /// the smoke-test assertion. Extracting the SQL ensures the smoke
-  /// test and the runtime gate stay in sync if this check ever
-  /// gains a third signal or moves to a different invariant.
-  pub fn schema_sentinel_satisfied(conn: &mut PgConnection) -> LemmyResult<bool> {
-    use diesel::sql_types::BigInt;
-    #[derive(diesel::QueryableByName)]
-    struct CountRow {
-      #[diesel(sql_type = BigInt)]
-      count: i64,
-    }
-    let row: CountRow = diesel::sql_query(
-      "SELECT \
-         (SELECT COUNT(*) FROM information_schema.tables \
-          WHERE table_schema = 'public' AND table_name = 'governance_log') \
-       + (SELECT COUNT(*) FROM information_schema.routines \
-          WHERE routine_schema = 'r' AND routine_name = 'parent_comment_ids') \
-       AS count",
-    )
-    .get_result::<CountRow>(conn)?;
-    Ok(row.count == 2)
-  }
-
-  pub fn apply_all_schema(conn: &mut PgConnection) -> LemmyResult<()> {
-    if schema_sentinel_satisfied(conn)? {
-      return Ok(());
-    }
-
-    // Sentinel missed → schema not fully populated by template
-    // restore. Delegate to the legacy bootstrap path so the
-    // migration + r-schema rebuild logic lives in exactly one place
-    // (CR finding #12 DRY).
-    apply_all_schema_legacy(conn).map_err(|e| anyhow::anyhow!("{e}"))?;
-    Ok(())
-  }
-
-  /// Tier 3 — pg_dump template captured once per nextest process (and
-  /// cached on disk across processes), then pg_restore'd into every
-  /// fresh container.
-  ///
-  /// **Throughput (PG18 + this laptop, nextest threads-required=4):**
-  /// template path 12m17s vs nextest legacy 12m02s for 67/67 e2e tests
-  /// — i.e. measurably **NEUTRAL**, not a 60× speedup as the original
-  /// plan projected. pg_restore (binary command replay) takes the same
-  /// wall-clock as Diesel `MigrationHarness::run_pending_migrations`
-  /// (SQL replay) on this hardware. Tier 3 ships as scaffolding for
-  /// future option-3 (per-test schema isolation) work, NOT as a
-  /// throughput improvement on this laptop. Likely speedup on slower
-  /// environments where Diesel migration runner overhead dominates.
-  /// Full discussion: `.claude/lessons/feedback_local_validation_cycle_2026_05_02.md`
-  /// Rule 4.
-  ///
-  /// Both `pg_dump` and `pg_restore` shell out via `docker exec` against
-  /// the live container, so no host-side Postgres client toolchain is
-  /// required. Container ships pg_dump 18.3 + pg_restore 18.3 (exact
-  /// version match for `pgautoupgrade:18-alpine`), eliminating the
-  /// version-skew risk.
-  pub mod pg_template {
-    use std::error::Error;
-    use std::path::PathBuf;
-    use tokio::process::Command;
-
-    /// In-process LazyLock for the dump bytes. Saves a disk read on
-    /// the second-and-subsequent test in the SAME nextest process.
-    /// Under nextest's process-per-test isolation this rarely fires
-    /// (each test = its own process), but it's cheap insurance and
-    /// load-bearing if anyone runs `cargo test` (one-process-many-tests).
-    static TEMPLATE_DUMP: tokio::sync::OnceCell<Vec<u8>> = tokio::sync::OnceCell::const_new();
-
-    /// Track whether `ensure_docker_cli()` has run successfully in
-    /// this process. Both the cold-bootstrap path AND the cache-hit
-    /// `pg_restore_into` path need a working `docker` binary, so the
-    /// preflight must run on whichever code path fires first (CR
-    /// finding #10). OnceCell ensures we pay the ~30ms `docker
-    /// --version` cost at most once per nextest process.
-    static DOCKER_CHECKED: tokio::sync::OnceCell<()> = tokio::sync::OnceCell::const_new();
-
-    /// Smallest plausible byte count for a healthy custom-format
-    /// pg_dump of the full Brehon schema. Used to reject:
-    /// - Truncated/half-written disk cache files in `load_or_build`
-    ///   (CR finding #18) — even a file starting with `PGDMP` could
-    ///   be corrupt and bypass the rebuild path.
-    /// - Implausibly empty bootstrap dumps in `template_dump_capture`
-    ///   (CR finding #15) — catches the "container produced no
-    ///   schema" failure mode.
-    ///
-    /// 1 KiB is well below any healthy dump (real dumps are
-    /// ~400 KiB+) but well above any truncation that would still
-    /// register as "looks like a file".
-    pub const MIN_TEMPLATE_DUMP_BYTES: usize = 1_024;
-
-    /// Compute the disk cache path. Lives under `target/tmp/` so
-    /// `cargo clean` clears it; suffixed with the test binary's mtime
-    /// (nanosecond precision) so any source-or-migration change
-    /// invalidates the cache (the `embed_migrations!` macro recompiles
-    /// on `migrations/` changes → new test binary → new mtime → cache
-    /// miss → rebuild).
-    ///
-    /// **Cache key precision (CR finding #3):** earlier draft used
-    /// `as_secs()` which collides on rebuilds within the same second
-    /// — incremental cargo can finish a touch-and-rebuild cycle under
-    /// 1s on a hot cache, producing a stale-dump-reuse hazard. Nanos
-    /// is overkill but cheap; collisions are now astronomically
-    /// unlikely (filesystem mtime resolution is the floor anyway).
-    fn cache_path() -> Result<PathBuf, Box<dyn Error>> {
-      let exe = std::env::current_exe()?;
-      let mtime = std::fs::metadata(&exe)?
-        .modified()?
-        .duration_since(std::time::UNIX_EPOCH)
-        .map_err(|e| -> Box<dyn Error> { format!("clock skew: {e}").into() })?
-        .as_nanos();
-      // Walk up from the e2e test binary
-      // (target/<profile>/deps/e2e-<hash>.exe) to find the workspace
-      // target directory: deps/ → debug/ or release/ → target/.
-      let target_root = exe
-        .parent()
-        .and_then(|p| p.parent())
-        .and_then(|p| p.parent())
-        .ok_or_else(|| -> Box<dyn Error> {
-          format!("cannot derive target root from {exe:?}").into()
-        })?;
-      Ok(
-        target_root
-          .join("tmp")
-          .join(format!("brehon-pg-template-{mtime}.dump")),
-      )
-    }
-
-    /// Lazily build (or return cached) bootstrap dump bytes. Cache
-    /// layers (fastest first):
-    ///   1. In-process `OnceCell` — same nextest process, second test.
-    ///   2. On-disk `target/tmp/...dump` — cross-process, keyed by test
-    ///      binary's mtime so source/migration changes invalidate.
-    ///   3. Cold bootstrap — ~30s container start + migrations + dump.
-    pub async fn ensure_template() -> Result<&'static Vec<u8>, Box<dyn Error>> {
-      TEMPLATE_DUMP.get_or_try_init(load_or_build).await
-    }
-
-    /// Pre-flight check that the `docker` CLI is on PATH. Tier 3
-    /// shells out via `docker exec` for pg_dump/pg_restore (rather
-    /// than using the bollard API like testcontainers-rs does
-    /// elsewhere), so the binary is a hard dependency. Surface a
-    /// clear actionable error here instead of letting `Command::new`
-    /// fail with `program not found` mid-run (CR findings #1, #10).
-    ///
-    /// OnceCell-gated so the ~30ms `docker --version` runs at most
-    /// once per nextest process, regardless of whether the first
-    /// caller is the cold-bootstrap path or a cache-hit
-    /// `pg_restore_into`.
-    pub(super) async fn ensure_docker_cli() -> Result<(), Box<dyn Error>> {
-      // Cloning the error path: OnceCell::get_or_try_init caches only
-      // success. On failure the next call retries, which is what we
-      // want — if the user installs docker after a missed first call,
-      // subsequent ensure_docker_cli() calls succeed.
-      DOCKER_CHECKED
-        .get_or_try_init(|| async {
-          let out = Command::new("docker")
-            .arg("--version")
-            .kill_on_drop(true)
-            .output()
-            .await
-            .map_err(|e| -> Box<dyn Error> {
-              format!(
-                "Tier 3 needs `docker` on PATH for pg_dump/pg_restore \
-                 via `docker exec`, but spawning failed: {e}. \
-                 Either install Docker CLI, or set \
-                 BREHON_E2E_NO_TEMPLATE=1 to bypass the template path \
-                 entirely (legacy migration runner)."
-              )
-              .into()
-            })?;
-          if !out.status.success() {
-            return Err::<(), Box<dyn Error>>(
-              format!(
-                "`docker --version` exited {:?}: {}. Tier 3 needs a \
-                 working docker CLI (or BREHON_E2E_NO_TEMPLATE=1).",
-                out.status.code(),
-                String::from_utf8_lossy(&out.stderr)
-              )
-              .into(),
-            );
-          }
-          Ok(())
-        })
-        .await
-        .map(|_| ())
-    }
-
-    async fn load_or_build() -> Result<Vec<u8>, Box<dyn Error>> {
-      let path = cache_path()?;
-      // Layer 2: disk cache hit. Validate the magic header AND the
-      // size before returning (CR findings #14 + #18). A truncated
-      // dump can still start with `PGDMP`, so the magic check alone
-      // isn't enough to prove integrity. Reuse `MIN_TEMPLATE_DUMP_BYTES`
-      // (the same floor `template_dump_capture` uses) to reject
-      // partially-written cache files. Either failure → delete +
-      // fall through to cold bootstrap (self-healing).
-      if let Ok(bytes) = tokio::fs::read(&path).await {
-        if bytes.starts_with(b"PGDMP") && bytes.len() >= MIN_TEMPLATE_DUMP_BYTES {
-          tracing::info!(
-            bytes = bytes.len(),
-            path = %path.display(),
-            "pg_template: disk cache hit (skipping bootstrap)"
-          );
-          return Ok(bytes);
-        }
-        tracing::warn!(
-          path = %path.display(),
-          bytes = bytes.len(),
-          first_bytes = ?bytes.get(..bytes.len().min(8)),
-          "pg_template: invalid disk cache (header or size below \
-           MIN_TEMPLATE_DUMP_BYTES), rebuilding"
-        );
-        let _ = tokio::fs::remove_file(&path).await;
-      }
-      // Layer 3: cold bootstrap (preflight docker first so the error
-      // surfaces before we boot a container that we can't dump).
-      ensure_docker_cli().await?;
-      let dump = build_template().await?;
-      // Best-effort write. Atomic via tmp-per-pid + rename. On race:
-      // - Linux/macOS rename atomically replaces the destination.
-      // - Windows rename FAILS with EEXIST when the destination already
-      //   exists (CR finding #4). On either OS, if a peer beat us to
-      //   write a fully-formed cache file, that file is correct and we
-      //   can simply discard our tmp and use the in-memory dump for
-      //   this process. Treating EEXIST/AlreadyExists as success keeps
-      //   the warning channel clean on Windows.
-      if let Some(parent) = path.parent() {
-        let _ = tokio::fs::create_dir_all(parent).await;
-      }
-      let tmp_path = path.with_extension(format!("dump.tmp.{}", std::process::id()));
-      if let Err(e) = tokio::fs::write(&tmp_path, &dump).await {
-        tracing::warn!(
-          error = %e,
-          tmp = %tmp_path.display(),
-          "pg_template: tmp write failed (continuing with in-memory dump)"
-        );
-      } else {
-        match tokio::fs::rename(&tmp_path, &path).await {
-          Ok(()) => {
-            tracing::info!(
-              bytes = dump.len(),
-              path = %path.display(),
-              "pg_template: disk cache populated"
-            );
-          }
-          Err(e) if path.exists() => {
-            // Peer already populated the cache. Our tmp is redundant.
-            // Covers the Windows EEXIST case AND the rare Linux race
-            // where a peer wrote between our `tokio::fs::read` miss
-            // above and our rename here.
-            let _ = tokio::fs::remove_file(&tmp_path).await;
-            tracing::debug!(
-              error = %e,
-              path = %path.display(),
-              "pg_template: peer populated cache between our read and rename — using in-memory dump"
-            );
-          }
-          Err(e) => {
-            // Genuine rename failure (permissions, cross-device,
-            // disk full). Keep the in-memory dump for this process;
-            // peers will pay their own bootstrap.
-            tracing::warn!(
-              error = %e,
-              tmp = %tmp_path.display(),
-              path = %path.display(),
-              "pg_template: rename to cache path failed (continuing with in-memory dump)"
-            );
-            let _ = tokio::fs::remove_file(&tmp_path).await;
-          }
-        }
-      }
-      Ok(dump)
-    }
-
-    async fn build_template() -> Result<Vec<u8>, Box<dyn Error>> {
-      use diesel::{Connection as _, PgConnection, RunQueryDsl, sql_types::Text};
-      let total = std::time::Instant::now();
-      let (container, host_port) = super::start_postgres_vanilla().await?;
-      let db_url = super::db_url(host_port);
-      let mut conn = PgConnection::establish(&db_url)
-        .map_err(|e| -> Box<dyn Error> { format!("template bootstrap establish: {e}").into() })?;
-      super::apply_all_schema_legacy(&mut conn)?;
-
-      // Query pg_extension for every non-builtin extension installed
-      // by the migrations, so the pg_dump command captures all of
-      // them dynamically (CR finding #17). The previous version hard-
-      // coded `pgcrypto`, `ltree`, `pg_trgm` — easy to forget when a
-      // new migration adds an extension. Excluding `plpgsql` (built-
-      // in to every Postgres database since 9.x; pg_dump skips it
-      // automatically and including it produces a benign warning).
-      #[derive(diesel::QueryableByName)]
-      struct ExtRow {
-        #[diesel(sql_type = Text)]
-        extname: String,
-      }
-      let ext_rows: Vec<ExtRow> = diesel::sql_query(
-        "SELECT extname FROM pg_extension \
-         WHERE extname <> 'plpgsql' \
-         ORDER BY extname",
-      )
-      .get_results(&mut conn)
-      .map_err(|e| -> Box<dyn Error> { format!("query pg_extension: {e}").into() })?;
-      let extensions: Vec<String> = ext_rows.into_iter().map(|r| r.extname).collect();
-      tracing::info!(
-        extensions = ?extensions,
-        "pg_template: extensions discovered for dump"
-      );
-
-      let dump = pg_dump(container.id(), &extensions).await?;
-      tracing::info!(
-        bytes = dump.len(),
-        ms = total.elapsed().as_millis(),
-        "pg_dump: template captured (bootstrap end-to-end)"
-      );
-      // container drops here; bootstrap done.
-      Ok(dump)
-    }
-
-    /// Hard timeout for pg_dump bootstrap. On healthy hardware the
-    /// dump completes in under 30s; 120s gives ~4× headroom for slow
-    /// laptops / contended Docker daemons. Without this timeout a
-    /// stalled docker exec would wedge the entire e2e suite with no
-    /// bounded failure (CR finding #7).
-    const PG_DUMP_TIMEOUT_SECS: u64 = 120;
-
-    /// Hard timeout for per-test pg_restore. Restore is faster than
-    /// pg_dump (no SQL parsing, just binary command replay); 90s is
-    /// ~6× the observed worst case.
-    const PG_RESTORE_TIMEOUT_SECS: u64 = 90;
-
-    /// Capture a custom-format pg_dump of `public` + `r` schemas + the
-    /// supplied extension list via `docker exec`. Custom format is more
-    /// compact than plain SQL and pg_restore's `--single-transaction
-    /// --exit-on-error` semantics give clean failure on per-test
-    /// restore.
-    ///
-    /// **Extensions:** `-e <name>` flags are required because
-    /// pg_restore `--clean --if-exists` drops the default `public`
-    /// schema, which cascades the extensions installed there. Without
-    /// these, restore fails with `type public.<X> does not exist` when
-    /// r.* functions reference extension-provided types (e.g.
-    /// `public.ltree`). The list is queried dynamically from
-    /// `pg_extension` in `build_template` (CR finding #17), so a
-    /// future migration that adds an extension automatically gets
-    /// captured without code changes here.
-    async fn pg_dump(container_id: &str, extensions: &[String]) -> Result<Vec<u8>, Box<dyn Error>> {
-      let start = std::time::Instant::now();
-      // Build the args list dynamically: fixed prefix + per-extension
-      // -e flags + format=custom suffix.
-      let mut args: Vec<String> = vec![
-        "exec".into(),
-        "-u".into(),
-        "postgres".into(),
-        container_id.into(),
-        "pg_dump".into(),
-        "-U".into(),
-        "lemmy".into(),
-        "-d".into(),
-        "lemmy".into(),
-        "--no-owner".into(),
-        "--no-privileges".into(),
-        "--schema=public".into(),
-        "--schema=r".into(),
-      ];
-      for ext in extensions {
-        args.push("-e".into());
-        args.push(ext.clone());
-      }
-      args.push("--format=custom".into());
-      let dump_future = Command::new("docker")
-        .args(&args)
-        .kill_on_drop(true)
-        .output();
-      let output = tokio::time::timeout(
-        std::time::Duration::from_secs(PG_DUMP_TIMEOUT_SECS),
-        dump_future,
-      )
-      .await
-      .map_err(|_e| -> Box<dyn Error> {
-        format!(
-          "pg_dump timed out after {PG_DUMP_TIMEOUT_SECS}s — \
-           docker daemon may be stalled. Check `docker ps` and consider \
-           BREHON_E2E_NO_TEMPLATE=1 as a workaround."
-        )
-        .into()
-      })??;
-      if !output.status.success() {
-        return Err(
-          format!(
-            "pg_dump failed (exit {:?}): {}",
-            output.status.code(),
-            String::from_utf8_lossy(&output.stderr)
-          )
-          .into(),
-        );
-      }
-      tracing::info!(
-        bytes = output.stdout.len(),
-        ms = start.elapsed().as_millis(),
-        "pg_dump: bootstrap dump captured"
-      );
-      Ok(output.stdout)
-    }
-
-    /// Restore the template dump into a fresh container via `docker
-    /// exec -i pg_restore` (dump bytes streamed on stdin). The fresh
-    /// container has POSTGRES_DB=lemmy already created (by entrypoint);
-    /// pg_restore loads schema into that empty DB.
-    ///
-    /// `--clean --if-exists` is paired with the dump's `-e` flags
-    /// (extensions captured) so DROP+CREATE for `public` + extensions
-    /// + tables + r.* objects all replay cleanly in one transaction.
-    pub async fn pg_restore_into(container_id: &str, dump: &[u8]) -> Result<(), Box<dyn Error>> {
-      use tokio::io::AsyncWriteExt;
-      // CR finding #10: cache-hit paths land here without going through
-      // load_or_build, so the docker CLI preflight must run here too.
-      // OnceCell makes it free on every call after the first.
-      ensure_docker_cli().await?;
-      let start = std::time::Instant::now();
-      let mut child = Command::new("docker")
-        .args([
-          "exec",
-          "-i",
-          "-u",
-          "postgres",
-          container_id,
-          "pg_restore",
-          "-U",
-          "lemmy",
-          "-d",
-          "lemmy",
-          "--no-owner",
-          "--no-privileges",
-          "--clean",
-          "--if-exists",
-          "--single-transaction",
-          "--exit-on-error",
-        ])
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        // kill_on_drop is essential for the timeout to actually
-        // terminate the underlying docker exec when we cancel the
-        // wait_with_output future. Without it, the OS process
-        // outlives the cancellation and ties up the docker daemon.
-        .kill_on_drop(true)
-        .spawn()?;
-      // Hard timeout (CR findings #7 + #11). Wraps the ENTIRE
-      // stdin-write + drop + wait flow because if the child becomes
-      // unresponsive while consuming stdin, the buffer fills and
-      // write_all stalls — we need that stall to count against the
-      // timeout. kill_on_drop on the Command above ensures the
-      // underlying docker exec is killed when this future is
-      // cancelled mid-write.
-      let output = match tokio::time::timeout(
-        std::time::Duration::from_secs(PG_RESTORE_TIMEOUT_SECS),
-        async {
-          {
-            let stdin = child
-              .stdin
-              .as_mut()
-              .ok_or_else(|| -> Box<dyn Error> { "no stdin on pg_restore child".into() })?;
-            stdin.write_all(dump).await?;
-            stdin.flush().await?;
-          }
-          // Drop stdin so pg_restore sees EOF.
-          drop(child.stdin.take());
-          child
-            .wait_with_output()
-            .await
-            .map_err(|e| -> Box<dyn Error> { format!("wait_with_output: {e}").into() })
-        },
-      )
-      .await
-      {
-        Ok(result) => result?,
-        Err(_) => {
-          return Err(
-            format!(
-              "pg_restore timed out after {PG_RESTORE_TIMEOUT_SECS}s — \
-               docker daemon or postgres may be stalled, or stdin \
-               write blocked because the child stopped consuming. \
-               Container ID: {container_id}. Consider \
-               BREHON_E2E_NO_TEMPLATE=1 as a workaround."
-            )
-            .into(),
-          );
-        }
-      };
-      if !output.status.success() {
-        return Err(
-          format!(
-            "pg_restore failed (exit {:?}): {}",
-            output.status.code(),
-            String::from_utf8_lossy(&output.stderr)
-          )
-          .into(),
-        );
-      }
-      tracing::info!(
-        ms = start.elapsed().as_millis(),
-        "pg_restore: template applied"
-      );
-      Ok(())
-    }
-  }
-
-  /// Bare container with no schema applied. Equivalent to the
-  /// historical `start_postgres` (pre-Tier-3). Used by:
-  ///   - `pg_template::build_template` for the bootstrap dump.
-  ///   - Tests that explicitly need to exercise cold migrations
-  ///     (`phase1_migrations_round_trip`,
-  ///     `v1_jm_a_backfill_populates_v0_snapshot`) — these pair the
-  ///     vanilla container with `schema_setup::run` to test the
-  ///     migration runner itself, which would no-op against a
-  ///     template-restored DB. CR finding #19.
-  pub async fn start_postgres_vanilla() -> Result<
-    (
-      testcontainers::ContainerAsync<testcontainers::GenericImage>,
-      u16,
-    ),
-    Box<dyn Error>,
-  > {
-    use testcontainers::{
-      GenericImage, ImageExt,
-      core::{IntoContainerPort, WaitFor},
-      runners::AsyncRunner,
-    };
-    let container = GenericImage::new("pgautoupgrade/pgautoupgrade", "18-alpine")
-      .with_exposed_port(5432.tcp())
-      .with_wait_for(WaitFor::message_on_stderr(
-        "database system is ready to accept connections",
-      ))
-      .with_env_var("POSTGRES_USER", "lemmy")
-      .with_env_var("POSTGRES_PASSWORD", "password")
-      .with_env_var("POSTGRES_DB", "lemmy")
-      .start()
-      .await?;
-    let host_port = container.get_host_port_ipv4(5432).await?;
-    Ok((container, host_port))
-  }
-
-  /// Legacy alias for the historical synchronous schema-apply path
-  /// — same body as the original `apply_all_schema`, kept for the
-  /// template-bootstrap path (where the sentinel deliberately misses)
-  /// and for `BREHON_E2E_NO_TEMPLATE=1` debug runs.
-  fn apply_all_schema_legacy(conn: &mut PgConnection) -> Result<(), Box<dyn Error>> {
-    conn.batch_execute("SELECT pg_advisory_lock(0);")?;
-    conn
-      .run_pending_migrations(MIGRATIONS)
-      .map_err(|e| -> Box<dyn Error> { format!("migrations failed: {e}").into() })?;
-    conn.batch_execute("DROP SCHEMA IF EXISTS r CASCADE; CREATE SCHEMA r;")?;
-    conn.batch_execute(include_str!(
-      "../../../crates/diesel_utils/replaceable_schema/utils.sql"
-    ))?;
-    conn.batch_execute(include_str!(
-      "../../../crates/diesel_utils/replaceable_schema/triggers.sql"
-    ))?;
-    Ok(())
-  }
-
-  /// Start a fresh `pgautoupgrade:18-alpine` container matching Lemmy's prod
-  /// image, returning the container handle (drop = teardown) and the
-  /// host-mapped port.
-  ///
-  /// **Tier 3 default path:** boots the container, then `pg_restore`s
-  /// the bootstrap template dump into it. The first caller in any
-  /// nextest process pays the ~30s bootstrap (vanilla container +
-  /// migrations + pg_dump → cached `Vec<u8>`); every other caller in
-  /// the same process pays only ~1-3s (fresh container + restore).
-  /// Existing call sites that pair this with `apply_all_schema(&mut
-  /// conn)` continue to compile and run unchanged — the sentinel in
-  /// `apply_all_schema` short-circuits when the restore already
-  /// populated the schema.
-  ///
-  /// **Legacy fallback:** set `BREHON_E2E_NO_TEMPLATE=1` to skip the
-  /// restore and rely on `apply_all_schema`'s legacy migration path.
-  /// Useful for debugging schema regressions where the dump might be
-  /// suspect.
-  pub async fn start_postgres() -> LemmyResult<(
-    testcontainers::ContainerAsync<testcontainers::GenericImage>,
-    u16,
-  )> {
-    let (container, host_port) = start_postgres_vanilla()
-      .await
-      .map_err(|e| anyhow::anyhow!("{e}"))?;
-    if std::env::var("BREHON_E2E_NO_TEMPLATE").as_deref() != Ok("1") {
-      let dump = pg_template::ensure_template()
-        .await
-        .map_err(|e| anyhow::anyhow!("{e}"))?;
-      pg_template::pg_restore_into(container.id(), dump)
-        .await
-        .map_err(|e| anyhow::anyhow!("{e}"))?;
-    }
-    Ok((container, host_port))
-  }
-
-  /// Build a standard test DB URL for the given mapped host port.
-  pub fn db_url(host_port: u16) -> String {
-    format!("postgres://lemmy:password@localhost:{host_port}/lemmy")
-  }
-
-  pub const SIGNING_SEED_HEX: &str =
-    "0000000000000000000000000000000000000000000000000000000000000001";
-
-  /// Spin a fresh Postgres, apply the full Brehon schema, build a real
-  /// `LemmyContext` wrapped in `Data`, and return the context handle,
-  /// the container guard (keep alive via `_container`), and the db URL.
-  pub async fn bootstrap() -> LemmyResult<(
-    testcontainers::ContainerAsync<testcontainers::GenericImage>,
-    Data<LemmyContext>,
-    String,
-  )> {
-    // SAFETY: tests run with --test-threads=1; no concurrent env mutation.
-    // These two env vars are intentionally process-scoped (NOT EnvVarGuard-wrapped):
-    // both are constant-valued ("1" / fixed signing seed) and bootstrap() has many
-    // callers across this test module — wrapping here would drop the guard at
-    // bootstrap() return, unsetting the var before the test body runs (see
-    // feedback_envvarguard_fixture_lifetime_footgun.md). LEMMY_DATABASE_URL IS
-    // guarded (per-call value) at the _g_db_url binding below.
-    unsafe {
-      std::env::set_var("LEMMY_INITIALIZE_WITH_DEFAULT_SETTINGS", "1");
-      std::env::set_var("GOVERNANCE_LOG_SIGNING_KEY", SIGNING_SEED_HEX);
-    }
-
-    let (container, host_port) = start_postgres().await?;
-    let db_url = db_url(host_port);
-    let _g_db_url = EnvVarGuard::set("LEMMY_DATABASE_URL", &db_url);
-
-    {
-      let mut sync_conn = PgConnection::establish(&db_url)?;
-      apply_all_schema(&mut sync_conn)?;
-    }
-
-    let pool: ActualDbPool = build_db_pool_for_tests();
-    let client = client_builder(&SETTINGS).build()?;
-    let middleware_client = ClientBuilder::new(client).build();
-    let secret = Secret {
-      id: 0,
-      jwt_secret: String::new().into(),
-    };
-    let rate_limit = RateLimit::with_debug_config();
-    let context = Data::new(LemmyContext::create(
-      pool,
-      middleware_client.clone(),
-      middleware_client,
-      secret,
-      rate_limit,
-    ));
-
-    Ok((container, context, db_url))
-  }
-
-  /// Seed a person/local_user pair, returning both the PersonId and the
-  /// `LocalUserView` callers need to invoke handlers.
-  pub async fn seed_user(
-    ctx: &LemmyContext,
-    instance_id: InstanceId,
-    name: &str,
-    is_admin: bool,
-  ) -> LemmyResult<(PersonId, LocalUserView)> {
-    let person_form = PersonInsertForm::test_form(instance_id, name);
-    let person = Person::create(&mut ctx.pool(), &person_form).await?;
-    let mut lu_form = if is_admin {
-      LocalUserInsertForm::test_form_admin(person.id)
-    } else {
-      LocalUserInsertForm::test_form(person.id)
-    };
-    lu_form.accepted_application = Some(true);
-    LocalUser::create(&mut ctx.pool(), &lu_form, vec![]).await?;
-    let view = LocalUserView::read_person(&mut ctx.pool(), person.id).await?;
-    Ok((person.id, view))
-  }
-
-  /// Seed a community for use in governance tests.
-  pub async fn seed_community(
-    ctx: &LemmyContext,
-    instance_id: InstanceId,
-  ) -> LemmyResult<Community> {
-    let community_form = CommunityInsertForm::new(
-      instance_id,
-      "testcomm".to_string(),
-      "Test Community".to_string(),
-      "comm-pubkey".to_string(),
-    );
-    Community::create(&mut ctx.pool(), &community_form).await
-  }
-
-  /// Seed `count` jurors and return their PersonIds in insertion order.
-  /// Names are `juror_<i>` zero-padded to ensure sort stability in lookups.
-  pub async fn seed_jurors(
-    ctx: &LemmyContext,
-    instance_id: InstanceId,
-    count: usize,
-  ) -> LemmyResult<Vec<PersonId>> {
-    let mut ids = Vec::with_capacity(count);
-    for i in 0..count {
-      let name = format!("juror_{i:02}");
-      let (pid, _) = seed_user(ctx, instance_id, &name, false).await?;
-      ids.push(pid);
-    }
-    Ok(ids)
-  }
-}
 
 #[tokio::test]
 async fn can_insert_moderation_case() -> lemmy_utils::error::LemmyResult<()> {
@@ -2502,13 +1711,6 @@ async fn modlog_view_returns_published_entries() -> lemmy_utils::error::LemmyRes
   );
 
   Ok(())
-}
-
-/// Tiny row shape for `sql_query` probes that return a single `id` column.
-#[derive(diesel::QueryableByName)]
-struct SingleI32 {
-  #[diesel(sql_type = diesel::sql_types::Int4)]
-  id: i32,
 }
 
 // ============================================================================
@@ -15490,73 +14692,8 @@ mod v1_sl_e_fixtures {
   }
 }
 
-mod v1_federation_inbound_a_fixtures {
-  //! **Process-env safety constraint:** every test in this module mutates
-  //! process-wide env vars (e.g. `LEMMY_DATABASE_URL`, `BREHON_DISABLE_*`).
-  //! Safety of those mutations is contingent on the Cargo runner flag
-  //! `--test-threads=1`. Running these tests with concurrent threads is
-  //! undefined behaviour and is forbidden — see the `EnvVarGuard` RAII guard.
-  use super::*;
-  use diesel::ExpressionMethods;
-  use diesel_async::{AsyncConnection, AsyncPgConnection, RunQueryDsl};
-  use lemmy_db_schema::source::governance::federation_peer::{
-    FederationPeerInsertForm, federation_inbox_check_peer_trust,
-  };
-  use lemmy_db_schema_file::InstanceId;
-  use lemmy_db_schema_file::enums::FederationPeerTrust;
-  use lemmy_db_schema_file::schema::{federation_peer, instance};
-  use lemmy_utils::error::LemmyResult;
-
-  async fn seed_federation_peer(
-    conn: &mut AsyncPgConnection,
-    domain: &str,
-    trust: FederationPeerTrust,
-  ) -> LemmyResult<InstanceId> {
-    let instance_id: InstanceId = diesel::insert_into(instance::table)
-      .values((
-        instance::domain.eq(domain),
-        instance::published_at.eq(diesel::dsl::now),
-      ))
-      .returning(instance::id)
-      .get_result(conn)
-      .await?;
-    let form = FederationPeerInsertForm {
-      instance_id,
-      trust_level: Some(trust),
-      added_by_actor: None,
-      notes: None,
-    };
-    diesel::insert_into(federation_peer::table)
-      .values(&form)
-      .execute(conn)
-      .await?;
-    Ok(instance_id)
-  }
-
-  #[tokio::test(flavor = "multi_thread")]
-  async fn federation_peer_trust_lookup_returns_seeded_state() -> LemmyResult<()> {
-    let (_container, _context, db_url) = governance_fixtures::bootstrap().await?;
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    let _instance_id = seed_federation_peer(
-      &mut conn,
-      "allowlisted.test",
-      FederationPeerTrust::Allowlisted,
-    )
-    .await?;
-    let trust = federation_inbox_check_peer_trust("allowlisted.test", &mut conn).await?;
-    assert_eq!(trust, FederationPeerTrust::Allowlisted);
-    Ok(())
-  }
-
-  #[tokio::test(flavor = "multi_thread")]
-  async fn federation_peer_trust_lookup_returns_unknown_for_first_seen() -> LemmyResult<()> {
-    let (_container, _context, db_url) = governance_fixtures::bootstrap().await?;
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    let trust = federation_inbox_check_peer_trust("unknown-peer.test", &mut conn).await?;
-    assert_eq!(trust, FederationPeerTrust::Unknown);
-    Ok(())
-  }
-}
+#[path = "e2e/federation_inbound_a.rs"]
+mod v1_federation_inbound_a_fixtures;
 
 mod v1_ship_2_fixtures {
   //! **Process-env safety constraint:** every test in this module mutates
@@ -16293,528 +15430,11 @@ async fn admin_audit_html_forbidden_for_non_admin() -> lemmy_utils::error::Lemmy
   Ok(())
 }
 
-mod v1_federation_inbound_b_fixtures {
-  //! **Process-env safety constraint:** every test in this module mutates
-  //! process-wide env vars (e.g. `LEMMY_DATABASE_URL`, `BREHON_DISABLE_*`).
-  //! Safety of those mutations is contingent on the Cargo runner flag
-  //! `--test-threads=1`. Running these tests with concurrent threads is
-  //! undefined behaviour and is forbidden — see the `EnvVarGuard` RAII guard.
-  use super::*;
-  use activitypub_federation::config::FederationConfig;
-  use activitypub_federation::traits::Activity as ActivityTrait;
-  use actix_web::error::ResponseError;
-  use actix_web::http::StatusCode;
-  use diesel::{ExpressionMethods, QueryDsl, SelectableHelper};
-  use diesel_async::{AsyncConnection, AsyncPgConnection, RunQueryDsl};
-  use lemmy_api_utils::context::LemmyContext;
-  use lemmy_apub_activities::protocol::governance::publish_label::PublishLabel;
-  use lemmy_apub_activities::protocol::governance::publish_sanction_notice::PublishSanctionNotice;
-  use lemmy_db_schema::source::governance::{
-    federation_inbox_nonce::FederationInboxNonceInsertForm,
-    federation_peer::FederationPeerInsertForm, remote_moderation_label::RemoteModerationLabel,
-    remote_sanction_notice::RemoteSanctionNotice,
-  };
-  use lemmy_db_schema_file::InstanceId;
-  use lemmy_db_schema_file::enums::FederationPeerTrust;
-  use lemmy_db_schema_file::schema::{
-    federation_inbox_dropped_log, federation_inbox_nonce, federation_peer, governance_config,
-    governance_log, instance, remote_moderation_label, remote_sanction_notice,
-  };
-  use lemmy_utils::error::{LemmyError, LemmyErrorType, LemmyResult};
-  use testcontainers::{ContainerAsync, GenericImage};
+#[path = "e2e/federation_inbound_b.rs"]
+mod v1_federation_inbound_b_fixtures;
 
-  async fn bootstrap_with_peer(
-    domain: &str,
-    trust: Option<FederationPeerTrust>,
-  ) -> LemmyResult<(
-    ContainerAsync<GenericImage>,
-    FederationConfig<LemmyContext>,
-    String,
-    InstanceId,
-  )> {
-    let (container, actix_context, db_url) = governance_fixtures::bootstrap().await?;
-    let federation_config = FederationConfig::builder()
-      .domain((**actix_context).settings().hostname.clone())
-      .app_data((**actix_context).clone())
-      .debug(true)
-      .http_fetch_limit(0)
-      .build()
-      .await?;
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    let peer_instance_id: i32 = diesel::insert_into(instance::table)
-      .values((
-        instance::domain.eq(domain),
-        instance::published_at.eq(diesel::dsl::now),
-      ))
-      .returning(instance::id)
-      .get_result(&mut conn)
-      .await?;
-    if let Some(t) = trust {
-      let form = FederationPeerInsertForm {
-        instance_id: InstanceId(peer_instance_id),
-        trust_level: Some(t),
-        added_by_actor: None,
-        notes: None,
-      };
-      diesel::insert_into(federation_peer::table)
-        .values(&form)
-        .execute(&mut conn)
-        .await?;
-    }
-    Ok((
-      container,
-      federation_config,
-      db_url,
-      InstanceId(peer_instance_id),
-    ))
-  }
-
-  #[tokio::test(flavor = "multi_thread")]
-  async fn blocklisted_peer_returns_403() -> LemmyResult<()> {
-    let (_container, fed_cfg, db_url, _peer_id) =
-      bootstrap_with_peer("blocked.test", Some(FederationPeerTrust::Blocklisted)).await?;
-    let context = fed_cfg.to_request_data();
-    let activity = build_minimal_sanction_notice_activity("blocked.test")?;
-    let result = ActivityTrait::receive(activity, &context).await;
-    assert!(result.is_err(), "wrapper must reject Blocklisted peer");
-    let err: LemmyError = result.err().unwrap();
-    assert!(matches!(
-      err.error_type,
-      LemmyErrorType::FederationPeerBlocklisted
-    ));
-    assert_eq!(err.status_code(), StatusCode::FORBIDDEN);
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    let drop_rows: i64 = federation_inbox_dropped_log::table
-      .filter(federation_inbox_dropped_log::source_instance.eq("blocked.test"))
-      .filter(federation_inbox_dropped_log::drop_reason.eq("blocklisted"))
-      .count()
-      .get_result(&mut conn)
-      .await?;
-    assert_eq!(drop_rows, 1);
-    Ok(())
-  }
-
-  #[tokio::test(flavor = "multi_thread")]
-  async fn per_peer_rate_limit_returns_429() -> LemmyResult<()> {
-    let (_container, fed_cfg, db_url, _peer_id) =
-      bootstrap_with_peer("rate-test.test", Some(FederationPeerTrust::Allowlisted)).await?;
-    let context = fed_cfg.to_request_data();
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    // governance_config is append-history: UPDATE mutates the existing seed row;
-    // the post-v1-federation-inbound-c reader (.order_by(valid_from.desc())) makes
-    // INSERT-with-newer-valid_from also safe. This test uses UPDATE for historical
-    // continuity; new override tests use INSERT (see appended_config_override_takes_effect_returns_429).
-    diesel::sql_query(
-      "UPDATE governance_config SET value_int = 2 \
-       WHERE scope = 'instance' AND key = 'federation.inbound.per_peer_rate_per_hour'",
-    )
-    .execute(&mut conn)
-    .await?;
-    for i in 0..2 {
-      let activity = build_unique_sanction_notice_activity("rate-test.test", i)?;
-      ActivityTrait::receive(activity, &context).await?;
-    }
-    let activity3 = build_unique_sanction_notice_activity("rate-test.test", 2)?;
-    let result = ActivityTrait::receive(activity3, &context).await;
-    assert!(result.is_err());
-    let err = result.err().unwrap();
-    assert!(matches!(
-      err.error_type,
-      LemmyErrorType::FederationPeerRateLimitExceeded
-    ));
-    assert_eq!(err.status_code(), StatusCode::TOO_MANY_REQUESTS);
-    Ok(())
-  }
-
-  #[tokio::test(flavor = "multi_thread")]
-  async fn appended_config_override_takes_effect_returns_429() -> LemmyResult<()> {
-    let (_container, fed_cfg, db_url, _peer_id) =
-      bootstrap_with_peer("override-test.test", Some(FederationPeerTrust::Allowlisted)).await?;
-    let context = fed_cfg.to_request_data();
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    // governance_config is append-only — INSERT a newer-valid_from row to
-    // override the seeded federation.inbound.per_peer_rate_per_hour (seed
-    // value_int = 100 per migration 2026-04-18-000000-0000_add_governance_config).
-    // Post-v1-federation-inbound-c, get_inbound_config_int reads the latest row;
-    // the override cap=2 means the 3rd activity should 429.
-    diesel::insert_into(governance_config::table)
-      .values((
-        governance_config::scope.eq("instance"),
-        governance_config::key.eq("federation.inbound.per_peer_rate_per_hour"),
-        governance_config::value_type.eq("int"),
-        governance_config::value_int.eq(Some(2_i64)),
-        governance_config::valid_from.eq(diesel::dsl::now),
-      ))
-      .execute(&mut conn)
-      .await?;
-    for i in 0..2 {
-      let activity = build_unique_sanction_notice_activity("override-test.test", i)?;
-      ActivityTrait::receive(activity, &context).await?;
-    }
-    let activity3 = build_unique_sanction_notice_activity("override-test.test", 2)?;
-    let result = ActivityTrait::receive(activity3, &context).await;
-    assert!(
-      result.is_err(),
-      "3rd activity must 429 against override cap=2"
-    );
-    let err = result.err().unwrap();
-    assert!(matches!(
-      err.error_type,
-      LemmyErrorType::FederationPeerRateLimitExceeded
-    ));
-    assert_eq!(err.status_code(), StatusCode::TOO_MANY_REQUESTS);
-    // Optional but recommended: assert the drop log row landed.
-    let drop_rows: i64 = federation_inbox_dropped_log::table
-      .filter(federation_inbox_dropped_log::source_instance.eq("override-test.test"))
-      .filter(federation_inbox_dropped_log::drop_reason.eq("rate_limit_peer"))
-      .count()
-      .get_result(&mut conn)
-      .await?;
-    assert_eq!(drop_rows, 1, "exactly one rate_limit_peer drop expected");
-    Ok(())
-  }
-
-  #[tokio::test(flavor = "multi_thread")]
-  async fn replayed_activity_returns_409() -> LemmyResult<()> {
-    let (_container, fed_cfg, db_url, _peer_id) =
-      bootstrap_with_peer("replay-test.test", Some(FederationPeerTrust::Allowlisted)).await?;
-    let context = fed_cfg.to_request_data();
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    let nonce_form = FederationInboxNonceInsertForm {
-      peer_instance: "replay-test.test".to_string(),
-      activity_id: "https://replay-test.test/activities/create/1".to_string(),
-    };
-    diesel::insert_into(federation_inbox_nonce::table)
-      .values(&nonce_form)
-      .execute(&mut conn)
-      .await?;
-    let activity = build_sanction_notice_with_id(
-      "replay-test.test",
-      "https://replay-test.test/activities/create/1",
-    )?;
-    let result = ActivityTrait::receive(activity, &context).await;
-    assert!(result.is_err());
-    let err = result.err().unwrap();
-    assert!(matches!(
-      err.error_type,
-      LemmyErrorType::FederationActivityReplayed
-    ));
-    assert_eq!(err.status_code(), StatusCode::CONFLICT);
-    Ok(())
-  }
-
-  #[tokio::test(flavor = "multi_thread")]
-  async fn allowlisted_happy_path_persists_advisory_row() -> LemmyResult<()> {
-    let (_container, fed_cfg, db_url, _peer_id) =
-      bootstrap_with_peer("happy.test", Some(FederationPeerTrust::Allowlisted)).await?;
-    let context = fed_cfg.to_request_data();
-    let activity = build_minimal_sanction_notice_activity("happy.test")?;
-    ActivityTrait::receive(activity, &context).await?;
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    let rows: i64 = remote_sanction_notice::table
-      .filter(remote_sanction_notice::source_instance.eq("happy.test"))
-      .count()
-      .get_result(&mut conn)
-      .await?;
-    assert_eq!(rows, 1);
-    let advisory: RemoteSanctionNotice = remote_sanction_notice::table
-      .filter(remote_sanction_notice::source_instance.eq("happy.test"))
-      .select(RemoteSanctionNotice::as_select())
-      .first(&mut conn)
-      .await?;
-    assert!(
-      advisory.local_case_id.is_none(),
-      "ADR-006: local_case_id MUST be NULL"
-    );
-    Ok(())
-  }
-
-  #[tokio::test(flavor = "multi_thread")]
-  async fn moderation_label_handler_persists_and_logs() -> LemmyResult<()> {
-    let (_container, fed_cfg, db_url, _peer_id) =
-      bootstrap_with_peer("label.test", Some(FederationPeerTrust::Allowlisted)).await?;
-    let context = fed_cfg.to_request_data();
-    let activity = build_minimal_publish_label_activity("label.test")?;
-    ActivityTrait::receive(activity, &context).await?;
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    let rows: i64 = remote_moderation_label::table
-      .filter(remote_moderation_label::source_instance.eq("label.test"))
-      .count()
-      .get_result(&mut conn)
-      .await?;
-    assert_eq!(rows, 1);
-    let label_row: RemoteModerationLabel = remote_moderation_label::table
-      .filter(remote_moderation_label::source_instance.eq("label.test"))
-      .select(RemoteModerationLabel::as_select())
-      .first(&mut conn)
-      .await?;
-    assert!(
-      label_row.local_case_id.is_none(),
-      "ADR-006: local_case_id MUST be NULL"
-    );
-    let log_count: i64 = governance_log::table
-      .filter(governance_log::entry_kind.eq("federation_label_received"))
-      .count()
-      .get_result(&mut conn)
-      .await?;
-    assert_eq!(log_count, 1);
-    Ok(())
-  }
-
-  fn build_minimal_sanction_notice_activity(
-    peer_domain: &str,
-  ) -> LemmyResult<PublishSanctionNotice> {
-    build_unique_sanction_notice_activity(peer_domain, 0)
-  }
-
-  fn build_unique_sanction_notice_activity(
-    peer_domain: &str,
-    seq: u32,
-  ) -> LemmyResult<PublishSanctionNotice> {
-    build_sanction_notice_with_id(
-      peer_domain,
-      &format!("https://{peer_domain}/activities/create/{seq}"),
-    )
-  }
-
-  fn build_sanction_notice_with_id(
-    peer_domain: &str,
-    activity_id: &str,
-  ) -> LemmyResult<PublishSanctionNotice> {
-    let actor_url = format!("https://{peer_domain}/u/admin");
-    let object_id = format!("https://{peer_domain}/objects/sanction/1");
-    let target_url = format!("https://{peer_domain}/u/target");
-    let val = serde_json::json!({
-      "type": "Create",
-      "actor": actor_url,
-      "to": ["https://www.w3.org/ns/activitystreams#Public"],
-      "cc": [],
-      "id": activity_id,
-      "object": {
-        "type": "SanctionNotice",
-        "id": object_id,
-        "actor": actor_url,
-        "target": target_url,
-        "action": "federation_quarantine_recommendation",
-        "scope": "federated_recommendation",
-        "summary": "test sanction notice",
-        "published": "2024-01-01T00:00:00Z"
-      }
-    });
-    let activity: PublishSanctionNotice = serde_json::from_value(val)?;
-    Ok(activity)
-  }
-
-  fn build_minimal_publish_label_activity(peer_domain: &str) -> LemmyResult<PublishLabel> {
-    let actor_url = format!("https://{peer_domain}/u/admin");
-    let object_id = format!("https://{peer_domain}/objects/label/1");
-    let target_url = format!("https://{peer_domain}/u/target");
-    let activity_url = format!("https://{peer_domain}/activities/create/1");
-    let val = serde_json::json!({
-      "type": "Create",
-      "actor": actor_url,
-      "to": ["https://www.w3.org/ns/activitystreams#Public"],
-      "cc": [],
-      "id": activity_url,
-      "object": {
-        "type": "ModerationLabel",
-        "id": object_id,
-        "actor": actor_url,
-        "target": target_url,
-        "label": "context-warning",
-        "published": "2024-01-01T00:00:00Z"
-      }
-    });
-    let activity: PublishLabel = serde_json::from_value(val)?;
-    Ok(activity)
-  }
-}
-
-mod v1_federation_inbound_e_fixtures {
-  //! **Process-env safety constraint:** every test in this module mutates
-  //! process-wide env vars (e.g. `LEMMY_DATABASE_URL`, `BREHON_DISABLE_*`).
-  //! Safety of those mutations is contingent on the Cargo runner flag
-  //! `--test-threads=1`. Running these tests with concurrent threads is
-  //! undefined behaviour and is forbidden — see the `EnvVarGuard` RAII guard.
-  use super::*;
-  use activitypub_federation::config::FederationConfig;
-  use activitypub_federation::traits::Activity as ActivityTrait;
-  use diesel::{ExpressionMethods, QueryDsl};
-  use diesel_async::{AsyncConnection, AsyncPgConnection, RunQueryDsl};
-  use lemmy_api_utils::context::LemmyContext;
-  use lemmy_apub_activities::protocol::governance::publish_sanction_notice::PublishSanctionNotice;
-  use lemmy_db_schema::source::governance::federation_peer::FederationPeerInsertForm;
-  use lemmy_db_schema_file::InstanceId;
-  use lemmy_db_schema_file::enums::FederationPeerTrust;
-  use lemmy_db_schema_file::schema::{
-    federation_inbox_dropped_log, federation_peer, governance_config, instance,
-    remote_sanction_notice,
-  };
-  use lemmy_utils::error::{LemmyErrorType, LemmyResult};
-  use testcontainers::{ContainerAsync, GenericImage};
-
-  async fn bootstrap_with_peer(
-    domain: &str,
-  ) -> LemmyResult<(
-    ContainerAsync<GenericImage>,
-    FederationConfig<LemmyContext>,
-    String,
-    InstanceId,
-  )> {
-    let (container, actix_context, db_url) = governance_fixtures::bootstrap().await?;
-    let federation_config = FederationConfig::builder()
-      .domain((**actix_context).settings().hostname.clone())
-      .app_data((**actix_context).clone())
-      .debug(true)
-      .http_fetch_limit(0)
-      .build()
-      .await?;
-    let mut conn = AsyncPgConnection::establish(&db_url).await?;
-    let peer_instance_id: i32 = diesel::insert_into(instance::table)
-      .values((
-        instance::domain.eq(domain),
-        instance::published_at.eq(diesel::dsl::now),
-      ))
-      .returning(instance::id)
-      .get_result(&mut conn)
-      .await?;
-    let form = FederationPeerInsertForm {
-      instance_id: InstanceId(peer_instance_id),
-      trust_level: Some(FederationPeerTrust::Allowlisted),
-      added_by_actor: None,
-      notes: None,
-    };
-    diesel::insert_into(federation_peer::table)
-      .values(&form)
-      .execute(&mut conn)
-      .await?;
-    Ok((container, federation_config, db_url, InstanceId(peer_instance_id)))
-  }
-
-  async fn seed_storage_cap(conn: &mut AsyncPgConnection, cap: i64) -> LemmyResult<()> {
-    diesel::insert_into(governance_config::table)
-      .values((
-        governance_config::scope.eq("instance"),
-        governance_config::key.eq("federation.inbound.per_peer_storage_cap"),
-        governance_config::value_type.eq("int"),
-        governance_config::value_int.eq(Some(cap)),
-        governance_config::valid_from.eq(diesel::dsl::now),
-      ))
-      .execute(conn)
-      .await?;
-    Ok(())
-  }
-
-  async fn preseed_sanction_notices(
-    conn: &mut AsyncPgConnection,
-    source_instance: &str,
-    n: usize,
-  ) -> LemmyResult<()> {
-    for i in 0..n {
-      diesel::sql_query(
-        "INSERT INTO remote_sanction_notice \
-         (source_instance, target_url, action, scope, summary, signature, published_at, received_at) \
-         VALUES ($1, $2, 'FederationQuarantineRecommendation'::sanction_action, \
-         'Instance'::sanction_scope, $3, $4, NOW(), NOW() - ($5 || ' seconds')::interval)",
-      )
-      .bind::<diesel::sql_types::Text, _>(source_instance)
-      .bind::<diesel::sql_types::Text, _>(format!("https://{source_instance}/target/{i}"))
-      .bind::<diesel::sql_types::Text, _>(format!("preseed sanction {i}"))
-      .bind::<diesel::sql_types::Text, _>(format!(
-        "https://{source_instance}/activity/preseed/{i}"
-      ))
-      .bind::<diesel::sql_types::Text, _>(format!("{}", 100 - i))
-      .execute(conn)
-      .await?;
-    }
-    Ok(())
-  }
-
-  fn build_sanction_notice_with_id(
-    peer_domain: &str,
-    activity_id: &str,
-  ) -> LemmyResult<PublishSanctionNotice> {
-    let actor_url = format!("https://{peer_domain}/u/admin");
-    let object_id = format!("https://{peer_domain}/objects/sanction/1");
-    let target_url = format!("https://{peer_domain}/u/target");
-    let val = serde_json::json!({
-      "type": "Create",
-      "actor": actor_url,
-      "to": ["https://www.w3.org/ns/activitystreams#Public"],
-      "cc": [],
-      "id": activity_id,
-      "object": {
-        "type": "SanctionNotice",
-        "id": object_id,
-        "actor": actor_url,
-        "target": target_url,
-        "action": "federation_quarantine_recommendation",
-        "scope": "federated_recommendation",
-        "summary": "test sanction notice",
-        "published": "2024-01-01T00:00:00Z"
-      }
-    });
-    let activity: PublishSanctionNotice = serde_json::from_value(val)?;
-    Ok(activity)
-  }
-
-  #[tokio::test(flavor = "multi_thread")]
-  async fn storage_cap_holds_under_concurrent_receivers() -> LemmyResult<()> {
-    let (_container, fed_cfg, db_url, _peer_id) =
-      bootstrap_with_peer("concurrent.test").await?;
-
-    let mut setup_conn = AsyncPgConnection::establish(&db_url).await?;
-    seed_storage_cap(&mut setup_conn, 5).await?;
-    preseed_sanction_notices(&mut setup_conn, "concurrent.test", 5).await?;
-
-    let context = fed_cfg.to_request_data();
-
-    let mut handles = Vec::with_capacity(8);
-    for i in 0..8u32 {
-      let context_i = context.reset_request_count();
-      let handle = tokio::spawn(async move {
-        let activity = build_sanction_notice_with_id(
-          "concurrent.test",
-          &format!("https://concurrent.test/activity/concurrent/{i}"),
-        )?;
-        ActivityTrait::receive(activity, &context_i).await
-      });
-      handles.push(handle);
-    }
-    for handle in handles {
-      handle
-        .await
-        .map_err(|e| LemmyErrorType::Unknown(format!("join: {e}")))?? ;
-    }
-
-    let mut verify_conn = AsyncPgConnection::establish(&db_url).await?;
-    let final_count: i64 = remote_sanction_notice::table
-      .filter(remote_sanction_notice::source_instance.eq("concurrent.test"))
-      .filter(remote_sanction_notice::admin_reviewed_at.is_null())
-      .count()
-      .get_result(&mut verify_conn)
-      .await?;
-    assert_eq!(
-      final_count,
-      5,
-      "per-peer storage cap must hold under concurrent receivers (Race B regression)"
-    );
-
-    let drop_log_count: i64 = federation_inbox_dropped_log::table
-      .filter(federation_inbox_dropped_log::source_instance.eq("concurrent.test"))
-      .filter(federation_inbox_dropped_log::drop_reason.eq("storage_cap_evicted"))
-      .count()
-      .get_result(&mut verify_conn)
-      .await?;
-    assert_eq!(
-      drop_log_count,
-      8,
-      "each concurrent receiver must fire exactly one eviction event (Race A regression)"
-    );
-
-    Ok(())
-  }
-}
+#[path = "e2e/federation_inbound_e.rs"]
+mod v1_federation_inbound_e_fixtures;
 
 mod v1_ship_3_fixtures {
   //! **Process-env safety constraint:** every test in this module mutates
